+++ scripts/winesapos-upgrade.sh
@@ -888,6 +888,11 @@ ${CMD_PACMAN} -Q gfs2-utils
 if [ $? -ne 0 ]; then
     ${CMD_YAY_INSTALL[*]} gfs2-utils
 fi
+
+${CMD_PACMAN} -Q glusterfs
+if [ $? -ne 0 ]; then
+    ${CMD_PACMAN_INSTALL[*]} glusterfs
+fi
 echo "Running 3.4.0 to 4.0.0 upgrades complete."
 
 echo "Upgrading system packages..."