+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - Add Distrobox for managing containers.
 - Change the build to use HTTP repositories for easier proxy caching.
 - Add support for the GFS2 file system.
+- Add support for the GlusterFS file system.
 
 ## 3.4.0 - The Game Mode Update